Fix Functions() method for ParenthesesExpr and update license headers
This should resolve opentofu/opentofu#3401

When provider-defined fnuctions were wrapped in parentheses, OpenTofu
failed to recognise them completely. This is because the ParenthesesExpr
embedded the Expression interface but didn't define it's own Value()
method and so it was not included in the `expression_funcs_gen` code.

This PR also updates license headers in expression_funcs_gen.go and its
generated output to be correct.

Signed-off-by: James Humphries <james@james-humphries.co.uk>